Refactor some code and fixed the PASS
