+++ Ircserv.cpp
@@ -6,7 +6,7 @@
 /*   By: bde-souz <bde-souz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/28 14:43:54 by bde-souz          #+#    #+#             */
-/*   Updated: 2025/03/09 20:09:43 by bde-souz         ###   ########.fr       */
+/*   Updated: 2025/03/10 11:55:08 by bde-souz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void Ircserv::createServer(const std::string& pass, unsigned int port)
 	_startTimer = std::time(0);
 	now = std::localtime(&_startTimer);
 
-	visualLoadingServer();
+	//visualLoadingServer();
 	
 	_password = pass;
 	_port = port;
@@ -199,7 +199,6 @@ void betterPrint(std::string str)
 	{
 		str.erase(str.size() - 1);
 	}
-	//std::cout << red << "Teste->" << str << "\n" << reset;
 }
 
 void Ircserv::bufferReader(int clientFd, char *buffer)
