+++ Makefile
@@ -15,8 +15,6 @@ SRC = main.cpp \
 	Ircserv.cpp\
 	debugFuncions.cpp \
 	helpFunctions.cpp \
-	channelControl.cpp\
-	utils.cpp \
 	commands/pass.cpp \
 	commands/ping.cpp \
 	commands/join.cpp \
@@ -29,6 +27,8 @@ SRC = main.cpp \
 	commands/kick.cpp \
 	commands/invite.cpp \
 	commands/unknown.cpp \
+	commands/part.cpp \
+	commands/topic.cpp \
 	signalHandler.cpp \
 
 
@@ -49,7 +49,7 @@ go:
 	
 govalgrind:
 	make;
-	valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes --tool=memcheck ./ircserv 6667 asd
+	valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes --tool=memcheck ./ircserv 6667 123
 clean: 
 	$(RM) $(RMFLAGS) $(OBJ)
 	$(RM) $(RMFLAGS) -r
