+++ commands/join.cpp
@@ -6,7 +6,7 @@
 /*   By: bde-souz <bde-souz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/11 15:40:35 by bde-souz          #+#    #+#             */
-/*   Updated: 2025/03/06 17:36:12 by bde-souz         ###   ########.fr       */
+/*   Updated: 2025/03/10 12:03:54 by bde-souz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -181,10 +181,9 @@ void Ircserv::commandJoin(const std::string &channel, const std::string &key)
 		std::string tempChannel = *it;
 		if (!commandJoinCheck(tempChannel))
 			return ;
-		//Se nao existir, cria um novo canal
 		if (!checkIfChannelExist(tempChannel))
 		{
-			std::cout << green << "Nao existe channel, criado um novo->" << tempChannel << "\n" << reset;
+			std::cout << green << "There is no channel, a new one has been created->" << tempChannel << "\n" << reset;
 			createNewChannel(tempChannel);
 			addClientToChannel(tempChannel, client);
 			changeClientToOperator(_clientFd, tempChannel);
