+++ commands/kick.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   kick.cpp                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: diodos-s <diodos-s@student.42.fr>          +#+  +:+       +#+        */
+/*   By: bde-souz <bde-souz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 13:26:55 by bde-souz          #+#    #+#             */
-/*   Updated: 2025/02/27 14:56:27 by diodos-s         ###   ########.fr       */
+/*   Updated: 2025/03/10 12:04:17 by bde-souz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,19 +14,19 @@
 
 void printAsciiValues(const std::string& str)
 {
-    for (std::string::const_iterator it = str.begin(); it != str.end(); ++it)
-    {
-        char c = *it;
-        std::cout << "Character: " << c << " ASCII: " << static_cast<int>(c) << std::endl;
-    }
+	for (std::string::const_iterator it = str.begin(); it != str.end(); ++it)
+	{
+		char c = *it;
+		std::cout << "Character: " << c << " ASCII: " << static_cast<int>(c) << std::endl;
+	}
 }
 
 void printVector(const std::vector<std::string>& vec)
 {
-    for (std::vector<std::string>::const_iterator it = vec.begin(); it != vec.end(); ++it)
-    {
-        std::cout << "vector->" << *it << std::endl;
-    }
+	for (std::vector<std::string>::const_iterator it = vec.begin(); it != vec.end(); ++it)
+	{
+		std::cout << "vector->" << *it << std::endl;
+	}
 }
