+++ signalHandler.cpp
@@ -6,7 +6,7 @@
 /*   By: bde-souz <bde-souz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 12:10:19 by bde-souz          #+#    #+#             */
-/*   Updated: 2025/03/05 18:39:49 by bde-souz         ###   ########.fr       */
+/*   Updated: 2025/03/10 12:02:08 by bde-souz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ void Ircserv::signalCatcher(void)
 void Ircserv::signalHandler(int signal)
 {
 	if(signal == SIGINT)
-		std::cout << red << "\nSigInt Ativado" << "\n" << reset;
+		std::cout << red << "\nCtrl + C pressed. Closing the server..." << "\n" << reset;
 	else if (signal == SIGQUIT)
-		std::cout << red << "\nSigQuit Ativado" << "\n" << reset;
+		std::cout << red << "\nCtrl + \\ pressed. Closing the server..." << "\n" << reset;
 	endServer = true;
 }
 
@@ -31,10 +31,7 @@ bool Ircserv::eofChecker(const std::string &buffer)
 {
 	printAsciiValues(buffer);
 	if (buffer.find('\04') != std::string::npos)
-	{
-		std::cout << red <<  "FOI ENCONTRADO NPOS" << "\n" << reset;
-		return true;
-	}
+		return (true);
 	else
 		return (false);
 }
@@ -63,30 +60,23 @@ std::string Ircserv::returnClientBuffer(const int &clientFd)
 	std::string tempStr = client.buffer;
 	client.buffer.clear();
 
-	//std::cout << "teste->" << tempStr << "\n";
 	return tempStr;
 }
 
 
 bool Ircserv::checkIfBufferHasEnd(const std::string &line)
 {
-	//printAsciiValues(line);
 	if (line.find(10) != std::string::npos)
 	{
 		Client &client = returnClientStructToModify(_clientFd);
-		//std::cout << red << "FOI ENCONTRADO O 10" << "\n" << reset;
 		client.bufferIsReady = true;
 		client.buffer += line;
-		//std::cout << "buffer atual :" << client.buffer << "<--" << "\n";
 		return (true);
 	}
 	else
 	{
-		std::cout << red << "NAO ENCONTROU" << "\n" << reset;
-		//std::cout << "CLIENTFD -> " << _clientFd << "\n";
 		Client &client = returnClientStructToModify(_clientFd);
 		client.buffer += line;
-		//std::cout << "buffer atual :" << client.buffer << "<--" << "\n";
 		return (false);
 	}
 }
